충돌í 해결

+++ src/main/java/shop/mtcoding/filmtalk/showtime/ShowtimeRepository.java
@@ -25,7 +25,7 @@ public interface ShowtimeRepository extends JpaRepository<Showtime,Long> {
     List<Showtime> mFindAll();
 
     @Query("select st from Showtime st left join fetch st.movie mt where st.id=:id")
-    Optional<Showtime> mFindById(@Param("id") Long id);
+    Optional<Showtime> mFindById(@Param("id") int id);
 
     @Query("select distinct st from Showtime st left join fetch st.movie mt where st.screen.id IN :ids ")
     List<Showtime> mFindByWithMovieScreenIds(@Param("ids") List<Long> ids);
@@ -35,12 +35,9 @@ public interface ShowtimeRepository extends JpaRepository<Showtime,Long> {
 
     @Query("SELECT s FROM Showtime s WHERE FUNCTION('DAY', s.startedAt) = :dateId AND s.movie.id = :movieId")
     List<Showtime> mFindByDateIdMovieId(@Param("dateId") Long dateId, @Param("movieId") Long movieId);
-
-
     @Query("SELECT st FROM Showtime st WHERE st.screen.id IN :screenIds AND CAST(st.startedAt AS date) = :showDate")
     List<Showtime> findByScreenIdsAndShowDate(
             @Param("screenIds") List<Long> screenIds,
             @Param("showDate") LocalDate showDate
     );
-
 }
